+++ packages/prisma/seed.ts
@@ -1,0 +1,14 @@
+import { logger } from "@packages/logger";
+import { prisma } from "@packages/prisma";
+
+(async () => {
+  // TODO: Seed database
+  logger.info("Seeding database...");
+})()
+  .then(() => {
+    prisma.$disconnect();
+  })
+  .catch((error) => {
+    prisma.$disconnect();
+    console.error(error);
+  });